✨ 캔버스 컴포넌트와 그리기 기능을 추가했습니다. 캔버스 그리는 로직을 훅으로 분리
