+++ src/components/canvas.tsx
@@ -2,6 +2,7 @@
 import React, { useRef, useEffect, useState } from "react";
 import FontButton from "./common/fontButton";
 import { disableBodyScroll, enableBodyScroll } from "body-scroll-lock";
+import useDrawCanvas from "@/hooks/useDrawCanvas";
 // style
 
 export const CANVAS_WIDTH = 300;
@@ -12,77 +13,17 @@ export default function Canvas({
 }: {
   canvasRef: React.RefObject<HTMLCanvasElement>;
 }) {
-  const [getCtx, setGetCtx] = useState<CanvasRenderingContext2D | null>(null);
-  const [painting, setPainting] = useState(false);
   const body = document.querySelector("body") as HTMLElement;
-  function getMousePos(
-    canvasDom: React.RefObject<HTMLCanvasElement>,
-    mouseEvent: React.TouchEvent<HTMLCanvasElement>
-  ) {
-    var rect = canvasDom.current?.getBoundingClientRect();
-    if (rect)
-      return {
-        x: mouseEvent.touches[0].clientX - (rect.left || 0),
-        y: mouseEvent.touches[0].clientY - (rect.top || 0),
-      };
-    return {
-      x: 0,
-      y: 0,
-    };
-  }
-
-  useEffect(() => {
-    const canvas = canvasRef.current;
-    if (canvas) {
-      canvas.width = CANVAS_WIDTH;
-      canvas.height = CANVAS_HEIGHT;
-      const ctx = canvas.getContext("2d");
-      if (ctx) {
-        ctx.lineJoin = "round";
-        ctx.lineWidth = 2.5;
-        ctx.strokeStyle = "#000000";
-        setGetCtx(ctx);
-      }
-    }
-  }, [canvasRef]);
-
-  const drawFn = (e: React.MouseEvent<HTMLCanvasElement, MouseEvent>) => {
-    // mouse position
-    const mouseX = e.nativeEvent.offsetX;
-    const mouseY = e.nativeEvent.offsetY;
-    // drawing
-    if (getCtx) {
-      if (!painting) {
-        getCtx.beginPath();
-        getCtx.moveTo(mouseX, mouseY);
-      } else {
-        getCtx.lineTo(mouseX, mouseY);
-        getCtx.stroke();
-      }
-    }
-  };
-
-  const drawMobileFn = (e: React.TouchEvent<HTMLCanvasElement>) => {
-    const { x, y } = getMousePos(canvasRef, e);
-    const mouseX = x;
-    const mouseY = y;
-    // drawing
-    if (getCtx) {
-      if (!painting) {
-        getCtx.beginPath();
-        getCtx.moveTo(mouseX, mouseY);
-      } else {
-        getCtx.lineTo(mouseX, mouseY);
-        getCtx.stroke();
-      }
-    }
-    setPainting(true);
-  };
-
-  const clearCanvas = () => {
-    setPainting(false);
-    getCtx?.clearRect(0, 0, CANVAS_WIDTH, CANVAS_HEIGHT);
-  };
+  const {
+    setPainting,
+    isDraw,
+    drawFn,
+    eraseFn,
+    drawMobileFn,
+    colorChange,
+    setIsDraw,
+    clearCanvas,
+  } = useDrawCanvas(canvasRef);
 
   return (
     <>
@@ -94,7 +35,7 @@ export default function Canvas({
             onMouseDown={() => setPainting(true)}
             onMouseUp={() => setPainting(false)}
             onMouseLeave={() => setPainting(false)}
-            onMouseMove={drawFn}
+            onMouseMove={isDraw ? drawFn : eraseFn}
             onTouchStart={(e) => {
               disableBodyScroll(body);
               setPainting(false);
@@ -109,6 +50,32 @@ export default function Canvas({
           ></canvas>
         </div>
       </div>
+      <button type="button" onClick={() => colorChange("#b7b7b7")}>
+        연필
+      </button>
+      <button type="button" onClick={() => colorChange("#010101")}>
+        검정펜
+      </button>
+      <button type="button" onClick={() => colorChange("#ff2a2a")}>
+        빨간펜
+      </button>
+      <button type="button" onClick={() => colorChange("#1582ff")}>
+        파란펜
+      </button>
+      <button type="button" onClick={() => colorChange("#00720d")}>
+        초록펜
+      </button>
+      <button type="button" onClick={() => colorChange("#ffffff")}>
+        하얀펜
+      </button>
+      <button
+        type="button"
+        onClick={() => {
+          setIsDraw(false);
+        }}
+      >
+        지우개
+      </button>
       <FontButton
         type="button"
         onClick={clearCanvas}
